backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-writable-nested==0.7.0
exceptiongroup==1.2.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
packaging==24.1
psycopg2==2.9.9
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
